[Kinetics] Properly type-cast variable to avoid compiler warnings

+++ src/kinetics/ReactionFactory.cpp
@@ -253,7 +253,7 @@ bool isThreeBody(const Reaction& R)
        if (trunc(reac.second) != reac.second) {
            return false;
        }
-       nreac += reac.second;
+       nreac += (size_t)reac.second;
     }
 
     // ensure that all products have integer stoichiometric coefficients
@@ -262,7 +262,7 @@ bool isThreeBody(const Reaction& R)
        if (trunc(prod.second) != prod.second) {
            return false;
        }
-       nprod += prod.second;
+       nprod += (size_t)prod.second;
     }
 
     // either reactant or product side involves exactly three species
